requirements.txt
torch>=2.2.0
transformers>=4.40.0
sentencepiece
pandas
pyyaml
scikit-learn
statsmodels
tqdm
requests
jupyter
